requirements.txt
pandas>=2.0
SQLAlchemy>=2.0
PyYAML>=6.0
pytest>=8.0
openpyxl>=3.1
